Move to package rename view variable
