+++ infullmvp-kotlin-koin/src/main/java/com/infullmobile/android/infullmvp/InFullMvpCustomView.kt
@@ -3,6 +3,6 @@ package com.infullmobile.android.infullmvp
 interface InFullMvpCustomView<out PresentedViewType : PresentedView<PresenterType>, PresenterType : InFullMvpPresenter<*>> {
 
     val presenter: PresenterType
-    val view: PresentedViewType
+    val presentedView: PresentedViewType
 
 }
