+++ sample-kotlin-koin/src/main/AndroidManifest.xml
@@ -4,7 +4,7 @@
     package="com.infullmobile.android.infullmvp.sample">
 
     <application
-        android:name=".SampleApplication"
+        android:name=".application.SampleApplication"
         android:allowBackup="false"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
