+++ sample-kotlin-koin/src/main/java/com/infullmobile/android/infullmvp/sample/activity/SampleKoinActivity.kt
@@ -7,6 +7,6 @@ import org.koin.core.parameter.parametersOf
 class SampleKoinActivity : InFullMvpActivity<SampleKoinView, SampleKoinPresenter>() {
 
     override val presenter: SampleKoinPresenter by inject { parametersOf(this) }
-    override val view: SampleKoinView by inject { parametersOf(this) }
+    override val presentedView: SampleKoinView by inject { parametersOf(this) }
 
 }
